Updates from Jayson
1.  We need to add "VSP" below the Hitachi Data Systems
2.  Need to change Hewlett Packard to just "HP Storage Works" so it
fits better on the page
3.  Add "XP48000" below HP Storage Works

For the main phone number, change 610-544-0571 to 877-210-8273

On the main page, for Hitachi HDS, Make it Hitachi Data Systems and
under it, list VSP and get rid of USPVM if you need space.

Make the onlinesales@allenmyland.com at the top a hot spot that takes
you to the contact us page.

On the home page, make the lins hotspots to go the corresponding pages.
 IE, you can click on services which will take you to the services
page.